+++ scripts/vlc.lua
@@ -1,0 +1,7 @@
+
+if (get_window_name()=="stream1") then
+    set_window_geometry2(279, 30, 630, 430);
+    undecorate_window();
+    set_on_top();
+    make_always_on_top();
+end