+++ CHANGELOG.md
@@ -1,3 +1,35 @@
+## [2.0.2-dev] - Attempted fix for OMXPlayer with OpenGL (Fake KMS) enabled
+
+Changed:
+* Added "--no-osd" command line switch to omxplayer command.  Per [this thread](https://www.raspberrypi.org/forums/viewtopic.php?t=159853), "omxplayer uses OpenVG for subtitles and status messsages which is not compatible with [the OpenGL (Fake KMS)] driver."
+
+## [2.0.1-dev] - Major OMX Bugfixes
+
+Changed:
+* App closing now spawns a new process to actually kill the OMX streams, it was getting cut off in the middle of closing everything due to it being an async process.
+* Wait for DOM to be shown before calling to start the streams--this was causing the Fullscreen on Resume problems.
+* Only start one stream at a time if we're in RotateStreams mode
+* Can jump to a specific stream in RotateStreams mode using notifications
+* Fixed and cleaned up all notifications and control from other modules
+* Fixed broken key bindings after MMM-KeyBindings upgrade
+* Fixed issue where module was trying to connect to PM2 while it was already connected (e.g. stop stream 1 and start stream 2 back to back).
+
+## [2.0.0] - Add VLC Streaming Support
+
+Added:
+
+* VLC Window Overlay support added. Use `localPlayer: 'vlc',` in your module configuration.
+* Module-wide debug option added for more verbose output: `debug: true,`
+
+Changed:
+
+* `shutdownDelay` parameter moved from the individual stream config sections to the main module config so it only has to be provided once. It has also changed from milliseconds to seconds.  Warning has been added if the timeout is less time than it takes to make it through the loop of streams (causes unnecessary restarts).
+* `hideFfmpegOutput` configuration option removed from stream config in favor of global `debug` module option.
+* Fixed bug where transition timer was not properly reset after module resume.
+* Added `hwAccel` stream option for `ffmpeg` to attempt to use hardware accelerated decoding. Encoding still uses CPU unfortunatly.
+    - You must update the node-rtsp-stream-es6 package too. This is most easily done by deleting your node_modules folder and re-running `npm install` on the module.
+* `RTSP-PLAY` notification now accepts an object `{ stream: "streamX", stopOthers: true }` which will stop other streams before starting the new stream.
+
 ## [1.2.2] - Auto-restart OMX Stream every X hours (Partially addresses #29)
 
 Changes: